pyproject.toml
[project]
name = "messedupsite"
version = "0.1.0"
description = "Messed Up Website - Django version"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.0",
    "psycopg2-binary>=2.9.0",
    "djangorestframework>=3.14.0",
    "Pillow>=12.2.0"
]

[dependency-groups]
dev = [
    "black>=26.3.0",
    "mypy>=1.20.0",
]

[tool.mypy]
ignore_missing_imports = true
